Move tests out of sequencing_primers_design.js
