+++ public/scripts/plugins/sequencing_primers/tests/test_expected_primers.js
@@ -1,0 +1,303 @@
+
+var expected863Primers = [{
+  name: 'Product 1 (forward)',
+  from: -100,
+  to: 399,
+  primer: {
+    name: 'Product 1 (forward) - primer',
+    from: -100,
+    to: -80,
+    sequence: 'AAAGGGAAAGGGAAACCCAAA',
+    meltingTemperature: 62.6,
+    gcContent: 0.429,
+  }
+},{
+  name: 'Product 2 (forward)',
+  from: 370,
+  to: 869,
+  primer: {
+    name: 'Product 2 (forward) - primer',
+    from: 370,
+    to: 399,
+    sequence: 'GTGTATCTATTCCTTTTATTGGAGAGGGAG',
+    gcContent: 0.4,
+    meltingTemperature: 64.4,
+  }
+},{
+  name: 'Product 3 (forward)',
+  from: 847,
+  to: 919,
+  primer: {
+    name: 'Product 3 (forward) - primer',
+    from: 847,
+    to: 869,
+    sequence: 'GGTATGCACGACATGCATTAGTT',
+    gcContent: 10/23,
+    meltingTemperature: 63.4,
+  }
+}];
+
+
+var expectedMikePrimers = [{
+  "name":"Product 1 (forward)",
+  "from": -148,
+  "to": 351,
+  "primer": {
+    "name":"Product 1 (forward) - primer",
+    "from": -148,
+    "to": -129,
+    "sequence": "TGCCACCTGACGTCTAAGAA",
+    "meltingTemperature": 63,
+    "gcContent": 0.5,
+  },
+},
+{
+  "name":"Product 2 (forward)",
+  "from": 293,
+  "to": 792,
+  "primer": {
+    "name":"Product 2 (forward) - primer",
+    "from": 293,
+    "to": 314,
+    "sequence": "CAAAATTGCTGTCTGCCAGGTG",
+    "meltingTemperature": 64.4,
+    "gcContent": 0.5,
+  },
+},
+{
+  "name":"Product 3 (forward)",
+  "from": 745,
+  "to": 1244,
+  "primer": {
+    "name":"Product 3 (forward) - primer",
+    "from": 745,
+    "to": 766,
+    "sequence": "TAACCTTTCATTCCCAGCGGTC",
+    "meltingTemperature": 64.4,
+    "gcContent": 0.5,
+  },
+},
+{
+  "name":"Product 4 (forward)",
+  "from": 1202,
+  "to": 1701,
+  "primer": {
+    "name":"Product 4 (forward) - primer",
+    "from": 1202,
+    "to": 1226,
+    "sequence": "GGGCTAGCAGGGAAAATAATGAATA",
+    "meltingTemperature": 62.9,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 5 (forward)",
+  "from": 1678,
+  "to": 2177,
+  "primer": {
+    "name":"Product 5 (forward) - primer",
+    "from": 1678,
+    "to": 1701,
+    "sequence": "TGTTCCATTATCAGGAGTGACATC",
+    "meltingTemperature": 62.2,
+    "gcContent": 0.4166666666666667,
+  },
+},
+{
+  "name":"Product 6 (forward)",
+  "from": 2147,
+  "to": 2646,
+  "primer": {
+    "name":"Product 6 (forward) - primer",
+    "from": 2147,
+    "to": 2171,
+    "sequence": "CAGCTAGATCGATACGCGAAAATTT",
+    "meltingTemperature": 63.5,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 7 (forward)",
+  "from": 2608,
+  "to": 3107,
+  "primer": {
+    "name":"Product 7 (forward) - primer",
+    "from": 2608,
+    "to": 2634,
+    "sequence": "CGAACAAACACGTTACTTAGAGGAAGA",
+    "meltingTemperature": 64.6,
+    "gcContent": 0.4074074074074074,
+  },
+},
+{
+  "name":"Product 8 (forward)",
+  "from": 3083,
+  "to": 3582,
+  "primer": {
+    "name":"Product 8 (forward) - primer",
+    "from": 3083,
+    "to": 3107,
+    "sequence": "CCGTAGGTGTCGTTAATCTTAGAGA",
+    "meltingTemperature": 63.1,
+    "gcContent": 0.44,
+  },
+},
+{
+  "name":"Product 9 (forward)",
+  "from": 3525,
+  "to": 4024,
+  "primer": {
+    "name":"Product 9 (forward) - primer",
+    "from": 3525,
+    "to": 3551,
+    "sequence": "CGTTACAAGTATTACTGTTAAGGAGCG",
+    "meltingTemperature": 63.4,
+    "gcContent": 0.4074074074074074,
+  },
+},
+{
+  "name":"Product 10 (forward)",
+  "from": 3990,
+  "to": 4489,
+  "primer": {
+    "name":"Product 10 (forward) - primer",
+    "from": 3990,
+    "to": 4016,
+    "sequence": "GGAAGAGTCTCGAGCAATTACTCAAAA",
+    "meltingTemperature": 64.8,
+    "gcContent": 0.4074074074074074,
+  },
+},
+{
+  "name":"Product 11 (forward)",
+  "from": 4454,
+  "to": 4953,
+  "primer": {
+    "name":"Product 11 (forward) - primer",
+    "from": 4454,
+    "to": 4478,
+    "sequence": "GGCGTGATTTTGTTTTACAAGGACA",
+    "meltingTemperature": 64.5,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 12 (forward)",
+  "from": 4883,
+  "to": 5382,
+  "primer": {
+    "name":"Product 12 (forward) - primer",
+    "from": 4883,
+    "to": 4906,
+    "sequence": "TGGTATTGTTGGAGCACCTATTAC",
+    "meltingTemperature": 62.5,
+    "gcContent": 0.4166666666666667,
+  },
+},
+{
+  "name":"Product 13 (forward)",
+  "from": 5331,
+  "to": 5830,
+  "primer": {
+    "name":"Product 13 (forward) - primer",
+    "from": 5331,
+    "to": 5355,
+    "sequence": "GAAACCAAAGAACGCTATGCAATTC",
+    "meltingTemperature": 63.3,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 14 (forward)",
+  "from": 5778,
+  "to": 6277,
+  "primer": {
+    "name":"Product 14 (forward) - primer",
+    "from": 5778,
+    "to": 5807,
+    "sequence": "GAGAGGGTATGACTGTCCATACTGAATATA",
+    "meltingTemperature": 64.7,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 15 (forward)",
+  "from": 6185,
+  "to": 6684,
+  "primer": {
+    "name":"Product 15 (forward) - primer",
+    "from": 6185,
+    "to": 6211,
+    "sequence": "GTTGGAGATTGGTTTGAGCATCAAATG",
+    "meltingTemperature": 65,
+    "gcContent": 0.4074074074074074,
+  },
+},
+{
+  "name":"Product 16 (forward)",
+  "from": 6665,
+  "to": 7164,
+  "primer": {
+    "name":"Product 16 (forward) - primer",
+    "from": 6665,
+    "to": 6684,
+    "sequence": "TATGCTCGGGCTCTTGATCC",
+    "meltingTemperature": 63.4,
+    "gcContent": 0.55,
+  },
+},
+{
+  "name":"Product 17 (forward)",
+  "from": 7134,
+  "to": 7633,
+  "primer": {
+    "name":"Product 17 (forward) - primer",
+    "from": 7134,
+    "to": 7158,
+    "sequence": "GAGACTGCTCATTGGATATTATCGA",
+    "meltingTemperature": 62.1,
+    "gcContent": 0.4,
+  },
+},
+{
+  "name":"Product 18 (forward)",
+  "from": 7613,
+  "to": 8112,
+  "primer": {
+    "name":"Product 18 (forward) - primer",
+    "from": 7613,
+    "to": 7633,
+    "sequence": "GCCGATGCTTTTGCATACGTA",
+    "meltingTemperature": 63.4,
+    "gcContent": 0.47619047619047616,
+  },
+},
+{
+  "name":"Product 19 (forward)",
+  "from": 8092,
+  "to": 8591,
+  "primer": {
+    "name":"Product 19 (forward) - primer",
+    "from": 8092,
+    "to": 8112,
+    "sequence": "TCTCATAGCTCACGCTGTAGG",
+    "meltingTemperature": 62.9,
+    "gcContent": 0.5238095238095238,
+  },
+},
+{
+  "name":"Product 20 (forward)",
+  "from": 8543,
+  "to": 8792,
+  "primer": {
+    "name":"Product 20 (forward) - primer",
+    "from": 8543,
+    "to": 8565,
+    "sequence": "CACGTTAAGGGATTTTGGTCATG",
+    "meltingTemperature": 62,
+    "gcContent": 0.43478260869565216,
+  }
+}];
+
+
+export default {expected863Primers, expectedMikePrimers};
