+++ public/scripts/plugins/sequencing_primers/tests/test_sequences.js
@@ -1,0 +1,147 @@
+
+var sequence863 = ('AAAAAAATGATTAAAAATTTATTGGCAATTTTAGATTTAAAATCTTTAG' +
+  'TACTCAATGCAATAAATTATTGGGGTCCTAAAAATAATAATGGCATACAGGGTGGTGATTTTGGTTACCC' +
+  'TATATCAGAAAAACAAATAGATACGTCTATTATAACTTCTACTCATCCTCGTTTAATTCCACATGATTTA' +
+  'ACAATTCCTCAAAATTTAGAAACTATTTTTACTACAACTCAAGTATTAACAAATAATACAGATTTACAAC' +
+  'AAAGTCAAACTGTTTCTTTTGCTAAAAAAACAACGACAACAACTTCAACTTCAACTACAAATGGTTGGAC' +
+  'AGAAGGTGGGAAAATTTCAGATACATTAGAAGAAAAAGTAA' + 
+  'GTGTATCTATTCCTTTTATTGGAGAGGGAG' +  // primer 2
+  'GAGGAAAAAACAGTACAACTATAGAAGCTAATTTTGCACATAACTCTAGTACTACTACTTTTCAACAGG' +
+  'CTTCAACTGATATAGAGTGGAATATTTCACAACCAGTATTGGTTCCCCCACGTAAACAAGTTGTAGCAAC' +
+  'ATTAGTTATTATGGGAGGTAATTTTACTATTCCTATGGATTTGATGACTACTATAGATTCTACAGAACAT' +
+  'TATAGTGGTTATCCAATATTAACATGGATATCGAGCCCCGATAATAGTTATAATGGTCCATTTATGAGTT' +
+  'GGTATTTTGCAAATTGGCCCAATTTACCATCGGGGTTTGGTCCTTTAAATTCAGATAATACGGTCACTTA' +
+  'TACAGGTTCTGTTGTAAGTCAAGTATCAGCTGGTGTATATGCCACTGTACGATTTGATCAATATGATATA' +
+  'CACAATTTAAGGACAATTGAAAAAACTT' + 
+  'GGTATGCACGACATGCATTAGTTA' + // primer 3
+  'TTATGGGAGGTAATTTTACTATTCCTATGGATTTGATGACTACTATAGA');
+
+var sequenceFromMike = ('TTATGACAACTTGACGGCTACATCATTCACTTTTTCTTCAC' +
+  'AACCGGCACGGAACTCGCTCGGGCTGGCCCCGGTGCATTTTTTAAATACCCGCGAGAAATAGAGTTGATC' +
+  'GTCAAAACCAACATTGCGACCGACGGTGGCGATAGGCATCCGGGTGGTGCTCAAAAGCAGCTTCGCCTGG' +
+  'CTGATACGTTGGTCCTCGCGCCAGCTTAAGACGCTAATCCCTAACTGCTGGCGGAAAAGATGTGACAGAC' +
+  'GCGACGGCGACAAGCAAACATGCTGTGCGACGCTGGCGATATCAAAATTGCTGTCTGCCAGGTGATCGCT' +
+  'GATGTACTGACAAGCCTCGCGTACCCGATTATCCATCGGTGGATGGAGCGACTCGTTAATCGCTTCCATG' +
+  'CGCCGCAGTAACAATTGCTCAAGCAGATTTATCGCCAGCAGCTCCGAATAGCGCCCTTCCCCTTGCCCGG' +
+  'CGTTAATGATTTGCCCAAACAGGTCGCTGAAATGCGGCTGGTGCGCTTCATCCGGGCGAAAGAACCCCGT' +
+  'ATTGGCAAATATTGACGGCCAGTTAAGCCATTCATGCCAGTAGGCGCGCGGACGAAAGTAAACCCACTGG' +
+  'TGATACCATTCGCGAGCCTCCGGATGACGACCGTAGTGATGAATCTCTCCTGGCGGGAACAGCAAAATAT' +
+  'CACCCGGTCGGCAAACAAATTCTCGTCCCTGATTTTTCACCACCCCCTGACCGCGAATGGTGAGATTGAG' +
+  'AATATAACCTTTCATTCCCAGCGGTCGGTCGATAAAAAAATCGAGATAACCGTTGGCCTCAATCGGCGTT' +
+  'AAACCCGCCACCAGATGGGCATTAAACGAGTATCCCGGCAGCAGGGGATCATTTTGCGCTTCAGCCATAC' +
+  'TTTTCATACTCCCGCCATTCAGAGAAGAAACCAATTGTCCATATTGCATCAGACATTGCCGTCACTGCGT' +
+  'CTTTTACTGGCTCTTCTCGCTAACCAAACCGGTAACCCCGCTTATTAAAAGCATTCTGTAACAAAGCGGG' +
+  'ACCAAAGCCATGACAAAAACGCGTAACAAAAGTGTCTATAATCACGGCAGAAAAGTCCACATTGATTATT' +
+  'TGCACGGCGTCACACTTTGCTATGCCATAGCATTTTTATCCATAAGATTAGCGGATCCTACCTGACGCTT' +
+  'TTTATCGCAACTCTCTACTGTTTCTCCATACCCGTTTTTTTGGGCTAGCAGGGAAAATAATGAATAAATG' +
+  'TATTCCAATGATAATTAATGGAATGATTCAAGATTTTGATAATTATGCATATAAAGAAGTTAAACTAAAT' +
+  'AATGATAATAGAGTAAAATTATCTGTCATTACTGAAAGTTCAGTTTCAAAAACATTAAATATCAAAGATA' +
+  'GAATTAATCTAAATTTAAATCAGATTGTGAATTTTTTATATACCGTTGGTCAACGATGGAAAAGTGAAGA' +
+  'ATATAATCGGCGACGAACCTATATCCGTGAGTTAAAAACATATCTTGGTTATTCTGATGAAATGGCAAGA' +
+  'TTAGAAGCGAATTGGATTGCAATGTTATTGTGCTCTAAAAGTGCTTTGTATGACATTGTTAATTATGATT' +
+  'TGGGCTCTATACACGTATTAGATGAATGGCTTCCACGTGGTGATTGCTATGTTAAAGCACAACCGAAAGG' +
+  'TGTTTCTGTTCACTTGTTAGCTGGTAATGTTCCATTATCAGGAGTGACATCTATTTTGCGTGCTATTTTA' +
+  'ACAAAAAATGAGTGCATTATTAAAACTTCGTCTTCAGATCCTTTTACTGCAAACGCTTTAGTTTCCAGTT' +
+  'TTATTGATGTTAATGCAGACCATCCAATAACCAAATCAATGTCTGTTATGTATTGGCCGCATGATGAAGA' +
+  'TATGACTCTATCTCAAAGAATAATGAATCATGCCGACGTGGTTATTGCTTGGGGTGGAGACGAGGCGATT' +
+  'AAATGGGCGGTAAAATATTCACCACCGCATGTCGATATTCTGAAATTTGGACCAAAGAAAAGCTTAAGTA' +
+  'TTATTGAAGCTCCTAAAGATATAGAAGCCGCAGCAATGGGGGTTGCTCATGATATTTGTTTCTATGACCA' +
+  'GCAAGCCTGCTTCTCTACTCAAGACGTTTATTATATAGGAGATAATTTACCTTTATTTTTAAATGAACTT' +
+  'GAAAAACAGCTAGATCGATACGCGAAAATTTTACCAAAAGGTTCAAATAGTTTTGATGAAAAAGCGGCGT' +
+  'TTACTCTTACTGAAAAAGAAAGTCTATTTGCTGGATATGAAGTGAGAAAGGGAGATAAGCAAGCTTGGTT' +
+  'AATAGTCGTATCACCTACAAATAGCTTTGGAAATCAACCGCTATCACGAAGTGTGTATGTTCATCAAGTA' +
+  'TCTGATATTAAAGAGATAATTCCTTTTGTTAATAAAAATAGAACACAAACTGTTTCTATTTATCCTTGGG' +
+  'AAGCGTCATTAAAATATCGAGATAAATTAGCAAGAAGTGGAGTTGAAAGAATTGTTGAATCAGGCATGAA' +
+  'TAATATTTTCAGAGTTGGAGGGGCTCATGATTCATTATCTCCTCTCCAGTACCTAGTTAGGTTTGTATCG' +
+  'CATGAGAGGCCATTTAATTATACGACAAAAGATGTTGCGGTTGAAATCGAACAAACACGTTACTTAGAGG' +
+  'AAGATAAGTTTTTAGTTTTTGTCCCATAGTTAAAGGAAATTATATGAAAGATGAAAGTGCTTTTTTTACG' +
+  'ATTGATCACATCATCAAGCTTGATAATGGTCAGTCTATCCGAGTTTGGGAAACACTCCCTAAAAAGAACG' +
+  'TACCAGAGAAAAAACATACAATACTTATTGCTTCGGGTTTTGCTAGAAGAATGGATCATTTTGCAGGTCT' +
+  'TGCTGAGTATTTATCTACTAACGGTTTTCATGTCATTCGCTACGATTCTTTGCATCATGTTGGATTAAGC' +
+  'AGTGGATGTATAAATGAATTTACGATGTCGATTGGAAAAAATAGCCTGCTTACAGTCGTAGATTGGCTTA' +
+  'CAGATCATGGTGTCGAACGAATAGGGCTGATTGCTGCTAGTTTGTCAGCGAGAATCGCCTATGAGGTAGT' +
+  'AAATAAAATTAAATTATCATTTTTAATTACGGCCGTAGGTGTCGTTAATCTTAGAGATACATTAGAAAAA' +
+  'GCATTGGAGTATGACTATTTGCAATTACCTATTTCAGAGTTACCAGAAGATCTTGACTTTGAAGGTCATA' +
+  'ATTTAGGATCGGAGGTCTTTGTTACAGATTGCTTTAAACATAATTGGGACACGTTAGACTCGACACTTAA' +
+  'TAGTGTTAAAGGATTAGCGATTCCATTTATTGCTTTTACTGCAAACGATGATAGTTGGGTAAAGCAAAGT' +
+  'GAAGTTATAGAGCTCATTGATAGCATTGAATCTAGTAATTGTAAGCTCTATTCGCTAATTGGAAGTTCAC' +
+  'ATGATCTTGGGGAAAATTTGGTTGTATTAAGAAATTTTTATCAATCAGTGACGAAGGCAGCCTTAGCATT' +
+  'AGATGATGGTTTATTGGATTTAGAGATAGACATTATTGAACCTCGATTTGAGGACGTTACAAGTATTACT' +
+  'GTTAAGGAGCGTAGATTAAAAAATGAAATTGAAAATGAATTATTAGAATTGGCTTAAATAAACAGAATCA' +
+  'CCAAAAAGGAATAGAGTATGAAGTTTGGAAATATTTGTTTTTCGTATCAACCACCAGGTGAAACTCATAA' +
+  'GCAAGTAATGGATCGCTTTGTTCGGCTTGGTATCGCCTCAGAAGAGGTAGGGTTTGATACATATTGGACC' +
+  'TTAGAACATCATTTTACAGAGTTTGGTCTTACGGGAAATTTATTTGTTGCTGCGGCTAACCTGTTAGGAA' +
+  'GAACTAAAACATTAAATGTTGGCACTATGGGGGTTGTTATTCCGACAGCACACCCAGTTCGACAGTTAGA' +
+  'AGACGTTTTATTATTAGATCAAATGTCGAAAGGTCGTTTTAATTTTGGAACCGTTCGAGGGCTATACCAT' +
+  'AAAGATTTTCGAGTATTTGGTGTTGATATGGAAGAGTCTCGAGCAATTACTCAAAATTTCTACCAGATGA' +
+  'TAATGGAAAGCTTACAGACAGGAACCATTAGCTCTGATAGTGATTACATTCAATTTCCTAAGGTTGATGT' +
+  'ATATCCCAAAGTGTACTCAAAAAATGTACCAACCTGTATGACTGCTGAGTCCGCAAGTACGACAGAATGG' +
+  'CTAGCAATACAAGGGCTACCAATGGTTCTTAGTTGGATTATTGGTACTAATGAAAAAAAAGCACAGATGG' +
+  'AACTCTATAATGAAATTGCGACAGAATATGGTCATGATATATCTAAAATAGATCATTGTATGACTTATAT' +
+  'TTGTTCTGTTGATGATGATGCACAAAAGGCGCAAGATGTTTGTCGGGAGTTTCTGAAAAATTGGTATGAC' +
+  'TCATATGTAAATGCGACCAATATCTTTAATGATAGCAATCAAACTCGTGGTTATGATTATCATAAAGGTC' +
+  'AATGGCGTGATTTTGTTTTACAAGGACATACAAACACCAATCGACGTGTTGATTATAGCAATGGTATTAA' +
+  'CCCTGTAGGCACTCCTGAGCAGTGTATTGAAATCATTCAACGTGATATTGATGCAACGGGTATTACAAAC' +
+  'ATTACATGCGGATTTGAAGCTAATGGAACTGAAGATGAAATAATTGCTTCCATGCGACGCTTTATGACAC' +
+  'AAGTCGCTCCTTTCTTAAAAGAACCTAAATAAATTACTTATTTGATACTAGAGATAATAAGGAACAAGTT' +
+  'ATGAAATTTGGATTATTTTTTCTAAACTTTCAGAAAGATGGAATAACATCTGAAGAAACGTTGGATAATA' +
+  'TGGTAAAGACTGTCACGTTAATTGATTCAACTAAATATCATTTTAATACTGCCTTTGTTAATGAACATCA' +
+  'CTTTTCAAAAAATGGTATTGTTGGAGCACCTATTACCGCAGCTGGTTTTTTATTAGGGTTAACAAATAAA' +
+  'TTACATATTGGTTCATTAAATCAAGTAATTACCACCCATCACCCTGTACGTGTAGCAGAAGAAGCCAGTT' +
+  'TATTAGATCAAATGTCAGAGGGACGCTTCATTCTTGGTTTTAGTGACTGCGAAAGTGATTTCGAAATGGA' +
+  'ATTTTTTAGACGTCATATCTCATCAAGGCAACAACAATTTGAAGCATGCTATGAAATAATTAATGACGCA' +
+  'TTAACTACAGGTTATTGTCATCCCCAAAACGACTTTTATGATTTTCCAAAGGTTTCAATTAATCCACACT' +
+  'GTTACAGTGAGAATGGACCTAAGCAATATGTATCCGCTACATCAAAAGAAGTCGTCATGTGGGCAGCGAA' +
+  'AAAGGCACTGCCTTTAACATTTAAGTGGGAGGATAATTTAGAAACCAAAGAACGCTATGCAATTCTATAT' +
+  'AATAAAACAGCACAACAATATGGTATTGATATTTCGGATGTTGATCATCAATTAACTGTAATTGCGAACT' +
+  'TAAATGCTGATAGAAGTACGGCTCAAGAAGAAGTGAGAGAATACTTAAAAGACTATATCACTGAAACTTA' +
+  'CCCTCAAATGGACAGAGATGAAAAAATTAACTGCATTATTGAAGAGAATGCAGTTGGCTCTCATGATGAC' +
+  'TATTATGAATCGACAAAATTAGCAGTGGAAAAAACAGGGTCTAAAAATATTTTATTATCCTTTGAATCAA' +
+  'TGTCCGATATTAAAGATGTAAAAGATATTATTGATATGTTGAACCAAAAAATCGAAATGAATTTACCATA' +
+  'ATAAAATTAAAGGCAATTTCTATATTAGATTGCCTTTTTAAATTTCTGTTGATATTAGGTATTACTGGAG' +
+  'AGGGTATGACTGTCCATACTGAATATAAAAGAAATCAAATCATTGCTAGTTCAGAAATTGATGATCTTAT' +
+  'CTTTATGACGAAACCACAAGAGTGGTCATTTGAAGAGCAAAAAGAAATACGGGATAAATTAGTTCGTGAG' +
+  'GCTTTTTATTTTCACTACAATAGAAATGAAGAATATAGAAATTATTGTATCAATCAGCATGTGAGTGATA' +
+  'ATTTACACACTATTGATGAAATACCCGTGTTTCCAACATCTGTTTTTAAATATAAGAAATTACATACTGT' +
+  'CACAGCCGAGGACATTGAAAATTGGTATACAAGTAGTGGAACTCGTGGAGTAAAAAGTCATATTGCACGT' +
+  'GATCGTCTTAGCATTGAACGCTTGCTTGGTTCTGTCAACTTCGGAATGAAATACGTTGGAGATTGGTTTG' +
+  'AGCATCAAATGGAATTGATAAATTTAGGACCAGATAGATTCAATACAAATAATATTTGGTTTAAATATGT' +
+  'CATGAGTTTGGTCGAGTTACTTTATCCGACTGAATTTACAGTTGATAATGACAAAATAGATTTTGAAAAA' +
+  'ACAGTAAAACATCTATTTAGAATTAAGAATAGTAAAAAAGACATTTGCTTAATTGGGCCACCATTTTTTG' +
+  'TGTATCTTTTGTGCCAATATATGAAAGAAAACAATATTGAATTTAAAGGAGGAGATAGAGTACATATTAT' +
+  'TACTGGTGGAGGATGGAAATCTAATCAGAATGACTCTTTAGATCGTGCTGATTTTAATCAATTATTAATG' +
+  'GATACTTTCCAACTCGACAAGATTAATCAAATTAGAGATACCTTTAATCAAGTTGAGCTTAATACTTGTT' +
+  'TTTTTGAAGATGAATTTCAAAGAAAACATGTTCCACCGTGGGTATATGCTCGGGCTCTTGATCCTGAAAC' +
+  'CTTGAAACCCGTAGCAGATGGTGAGATCGGGTTGTTAAGTTATATGGACGCCTCATCAACTGCTTACCCT' +
+  'GCTTTTATTGTTACTGATGATATCGGTATTGTAAAAGAAATTAGAGAACCAGATCCTTACCCAGGGGTAA' +
+  'CTGTTGAGATTGTTCGGCGCTTAAATACACGTGCGCAAAAAGGATGCGCGCTCTCTATGGCTAATGTCAT' +
+  'ACAAAAGAATATCAAGGATTAAGTTATGATTGTTGATGGTAGAGTTTCAAAAATAGTATTAGCGTCAATA' +
+  'AGAAATAATATATATAAGGTATTTATTACTGTAAATTCACCAATAAAGTTCATCGCTGGACAATTTGTAA' +
+  'TGGTCACGATTAATGGGAAAAAATGCCCTTTTTCAATTGCGAATTGCCCGACAAAAAATTACGAAATAGA' +
+  'ATTGCATATTGGTAGTTCGAATAGAGACTGCTCATTGGATATTATCGAATATTTTGTCGATGCTCTTGTT' +
+  'GAGGAAGTCGCAATTGAGTTAGATGCTCCCCATGGAAACGCTTGGTTACGGTCTGAAAGTAATAACCCAT' +
+  'TGCTATTAATTGCGGGAGGTACAGGTTTATCATATATAAATAGCATTCTAACAAATTGCTTAAATAGGAA' +
+  'TATACCTCAAGATATTTATCTTTACTGGGGAGTAAAAGACAGTTCTCTTTTGTATGAAGATGAAGAGTTA' +
+  'CTAAACTTATCACTAAACAACAAAAACTTTCATTATATTCCTGTTATTGAAGATAAAAGTGAAGAATGGA' +
+  'TAGGGAGAAAAGGCACTGTTCTTGATGCTGTCATGGAAGATTTTACTGATCTTACTTATTTTGATATTTA' +
+  'TGTTTGTGGACCCTTCATGATGGCTAAAACAGCAAAAGAAAAATTAATTGAAGAGAAAAAAGCAAAGTCA' +
+  'GAACAGATGTTTGCCGATGCTTTTGCATACGTATAAGAGATGAGACCTCCGGCAAAAAAACGGGCAAGGT' +
+  'GTCACCACCCTGCCCTTTTTCTTTAAAACCGAAAAGATTACTTCGCGTTCTGCAGGCTTCCTCGCTCACT' +
+  'GACTCGCTGCGCTCGGTCGTTCGGCTGCGGCGACCGGTATCAGCTCACTCAAAGGCGGTAATACGGTTAT' +
+  'CCACAGAATCAGGGGATAACGCAGGAAAGAACATGTGAGCAAAAGGCCAGCAAAAGGCCAGGAACCGTAA' +
+  'AAAGGCCGCGTTGCTGGCGTTTTTCCATAGGCTCCGCCCCCCTGACGAGCATCACAAAAATCGACGCTCA' +
+  'AGTCAGAGGTGGCGAAACCCGACAGGACTATAAAGATACCAGGCGTTTCCCCCTGGAAGCTCCCTCGTGC' +
+  'GCTCTCCTGTTCCGACCCTGCCGCTTACCGGATACCTGTCCGCCTTTCTCCCTTCGGGAAGCGTGGCGCT' +
+  'TTCTCATAGCTCACGCTGTAGGTATCTCAGTTCGGTGTAGGTCGTTCGCTCCAAGCTGGGCTGTGTGCAC' +
+  'GAACCCCCCGTTCAGCCCGACCGCTGCGCCTTATCCGGTAACTATCGTCTTGAGTCCAACCCGGTAAGAC' +
+  'ACGACTTATCGCCACTGGCAGCAGCCACTGGTAACAGGATTAGCAGAGCGAGGTATGTAGGCGGTGCTAC' +
+  'AGAGTTCTTGAAGTGGTGGCCTAACTACGGCTACACTAGAAGGACAGTATTTGGTATCTGCGCTCTGCTG' +
+  'AAGCCAGTTACCTTCGGAAAAAGAGTTGGTAGCTCTTGATCCGGCAAACAAACCACCGCTGGTAGCGGTG' +
+  'GTTTTTTTGTTTGCAAGCAGCAGATTACGCGCAGAAAAAAAGGATCTCAAGAAGATCCTTTGATCTTTTC' +
+  'TACGGGGTCTGACGCTCAGTGGAACGAAAACTCACGTTAAGGGATTTTGGTCATGAGATTATCAAAAAGG' +
+  'ATCTTCACCTAGATCCTTTTAAATTAAAAATGAAGTTTTAAATCAATCTAAAGTATATATGAGTAAACTT' +
+  'GGTCTGACAGCTCGAGGCTTGGATTCTCACCAATAAAAAACGCCCGGCGGCAACCGAGCGTTCTGAACAA' +
+  'ATCCAGATGGAGTTCTGAGGTCATTACTGGATCTATCAACAGGAGTCCAAGCGAGCTCGATATCAAATTA' +
+  'CG');
+
+export default {sequence863, sequenceFromMike};
